refactor(ui): extract PaneCommandExecutor
